% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_motif_logo.R
\name{gg_motif_logos}
\alias{gg_motif_logos}
\title{Several sequence logos in one figure}
\usage{
gg_motif_logos(x, motifs = NULL, ncol = 1, nrow = NULL, shared_y = TRUE, ...)
}
\arguments{
\item{x}{A [meme_result].}

\item{motifs}{Motif labels to draw; defaults to all.}

\item{ncol, nrow}{Grid layout, passed to [patchwork::wrap_plots()].}

\item{shared_y}{Give every panel the same y axis limit, so panels are
directly comparable.}

\item{...}{Passed to [gg_motif_logo()].}
}
\value{
A `patchwork` object.
}
\description{
Several sequence logos in one figure
}
\examples{
res <- read_meme(system.file("extdata", "example_meme.xml", package = "memeplotr"))
gg_motif_logos(res, ncol = 2)
}
