% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_tree_motifs.R
\name{gg_tree_panel}
\alias{gg_tree_panel}
\title{Compose a tree panel with any tip-ordered plot}
\usage{
gg_tree_panel(tree_plot, panel, widths = c(1, 2.4), guides = "collect", ...)
}
\arguments{
\item{tree_plot}{A `ggplot` from [gg_phylo()].}

\item{panel}{A `ggplot` to place on the right, or a list of them.}

\item{widths}{Relative panel widths, recycled over
`1 + length(panel)` panels.}

\item{guides}{Passed to [patchwork::plot_layout()]; `"collect"` merges the
legends of all panels.}

\item{...}{Further arguments for [patchwork::plot_layout()].}
}
\value{
A `patchwork` object.
}
\description{
Places a [gg_phylo()] panel to the left of another plot and aligns the two
vertically. The right-hand plot must put tip *i* of
[tree_tip_order()] at y = `n - i + 1` and use the same `expand_y`, which is
what [gg_motif_map()], [gg_motif_heatmap()] and [gg_motif_counts()] do when
given `order = "tree"`.
}
\examples{
res <- read_meme(system.file("extdata", "example_meme.xml", package = "memeplotr"))
nwk <- system.file("extdata", "example_tree.nwk", package = "memeplotr")
gg_tree_panel(
  gg_phylo(nwk),
  gg_motif_map(res, order = "tree", tree = nwk, keep_empty = TRUE) +
    ggplot2::theme(axis.text.y = ggplot2::element_blank())
)
}
\seealso{
[gg_tree_motifs()]
}
