% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meme_result-methods.R
\name{meme_result}
\alias{meme_result}
\alias{print.meme_result}
\alias{summary.meme_result}
\alias{as.data.frame.meme_result}
\title{Methods for `meme_result` objects}
\usage{
\method{print}{meme_result}(x, ...)

\method{summary}{meme_result}(object, ...)

\method{as.data.frame}{meme_result}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{A `meme_result` from [read_meme()].}

\item{...}{Passed on to methods.}

\item{object}{A `meme_result` from [read_meme()].}

\item{row.names, optional}{Ignored; present for S3 consistency.}
}
\value{
`print()` returns `x` invisibly; `summary()` returns a tibble with one
  row per motif plus occupancy statistics; `as.data.frame()` returns the site
  table as a plain data frame.
}
\description{
Methods for `meme_result` objects
}
\examples{
res <- read_meme(system.file("extdata", "example_meme.xml", package = "memeplotr"))
print(res)
summary(res)
head(as.data.frame(res))

}
