% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_phylo.R
\name{read_tree}
\alias{read_tree}
\title{Read a phylogeny}
\usage{
read_tree(x)
}
\arguments{
\item{x}{An `ape::phylo` object, a path to a Newick file, or a Newick string
(recognised by the trailing `;`).}
}
\value{
An `ape::phylo` object.
}
\description{
Thin wrapper around [ape::read.tree()] that also accepts an existing `phylo`
object, a Newick string, or a file path, so that every memeplotr function
can take `tree =` in whichever form you have.
}
\examples{
read_tree(system.file("extdata", "example_tree.nwk", package = "memeplotr"))
read_tree("((a:1,b:1):1,c:2);")
}
